third_party/blink/web_tests/platform/generic/http/tests/websocket/fragmented-frames-expected.txt

Receive fragmented WebSocket frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onmessage() was called. (message = "First message")
onmessage() was called. (message = "Second message")
onmessage() was called. (message = "Third message")
PASS closeEvent.wasClean is true
PASS actualMessages.length === expectedMessages.length is true
PASS actualMessages[0] is "First message"
PASS actualMessages[1] is "Second message"
PASS actualMessages[2] is "Third message"
PASS successfullyParsed is true

TEST COMPLETE
